reorg plc constants, bring plc functionality into lib/identity
